Move diagnostic management out of the worktree crate
