+++ crates/worktree/Cargo.toml
@@ -36,7 +36,6 @@ ignore.workspace = true
 itertools.workspace = true
 language.workspace = true
 log.workspace = true
-lsp.workspace = true
 parking_lot.workspace = true
 postage.workspace = true
 rpc.workspace = true
